+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library("testthat")
+library("quanteda")
 library("quanteda.textmodels")
 
 test_check("quanteda.textmodels")