+++ NEWS.md
@@ -1,3 +1,8 @@
+# quanteda.textmodels 0.9.3
+
+* Fixed `textmodel_svm()` to default to the "L2-regularized L2-loss support vector classification (dual)" model, and added the `type` argument to allow this to be overridden. (#45)
+* Loosened a test tolerance that was causing a text to fail on the macM1 platform.
+
 # quanteda.textmodels 0.9.2
 
 * Added a logistic regression classifier, `textmodel_lr()`.